fraud_detection/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pydantic-settings
bcrypt
python-jose
joblib
scikit-learn
xgboost
numpy
pandas
email-validator
